requirements.txt
faiss-cpu>=1.7.4
sentence-transformers>=2.7.0
numpy>=1.24.0
tqdm>=4.65.0
torch>=2.0.0